refactor(dna): modify get_notifications_for_agents to reduce number of get_links (1 for each mew by agent + 1)
