+++ dnas/mewsfeed/zomes/coordinator/mews/src/lib.rs
@@ -1,5 +1,6 @@
 use hdk::prelude::*;
 pub mod agent_mews;
+pub mod agent_to_notifications;
 pub mod all_mews;
 pub mod cashtag_to_mews;
 pub mod followed_creators_mews;
@@ -9,7 +10,6 @@ pub mod mention_to_mews;
 pub mod mew;
 pub mod mew_to_responses;
 pub mod mew_with_context;
-pub mod notifications;
 pub mod pinner_to_mews;
 pub mod random_mews;
 pub mod search_tags;
